+++ CHANGELOG.md
@@ -1,10 +1,45 @@
-
 Change Log
 ==========
 
 Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
+## [0.2.2] - 2019-05-20
+
+Changed:
+
+  * Add simple page cropping processor crop
+  * Respect border cropping in segment_word
+  * Add parameter overwrite_words in recognize
+  * Make higher TextEquivs consistent after recognize
+  
+Fixed:
+
+  * Remove invalid @externalRef from MetadataItem
+  * Retain pageId in output (i.e. link to structMap)
+
+## [0.2.1] - 2019-02-28
+
+Fixed:
+
+  * workspace.add_file was wrong in segment_word
+
+## [0.2.0] - 2019-02-28
+
+Changed:
+
+  * Adapt to OCR-D/core 1.0.0b5 API
+
+## [0.1.3] - 2019-01-04
+
+Fixed:
+
+  * Override locale to POSIX before importing tesserocr
+
+Changed:
+
+  * split recognizing existing glyphs vs. all in word
 
 ## [0.1.2] - 2018-09-03
 
@@ -27,6 +62,7 @@ Changed:
   * Recognition with proper support for textequiv_level, drop `page` level
 
 <!-- link-labels -->
+[0.1.3]: v0.1.3...v0.1.2
 [0.1.2]: v0.1.2...v0.1.1
 [0.1.1]: v0.1.1...v0.1.0
 [0.1.0]: ../../compare/HEAD...v0.1.0